# docs/swagger.yaml
basePath: /api
definitions:
  dto.PostDTO:
    properties:
      content:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      title:
        type: string
      userId:
        type: integer
    type: object
  dto.UserDTO:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  request.CommentReq:
    properties:
      content:
        type: string
      postId:
        type: integer
      userId:
        type: integer
    type: object
  response.ResultResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
      success:
        type: boolean
    type: object
host: localhost:9080
info:
  contact: {}
  description: 这是一个基于 Gin 的 个人博客系统后端API
  title: 个人博客系统后端 API 文档
  version: "1.0"
paths:
  /login:
    post:
      description: 用户通过用户名和密码登录
      parameters:
      - description: 用户登录参数
        in: body
        name: UserDTO
        required: true
        schema:
          $ref: '#/definitions/dto.UserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResultResponse'
      summary: 用户登录
      tags:
      - 登录
  /login/logout:
    post:
      description: 用户退出登录
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResultResponse'
      summary: 用户退出登录
      tags:
      - 登录
  /login/register:
    post:
      description: 注册一个用户
      parameters:
      - in: formData
        name: authenticationFlag
        type: string
      - in: formData
        name: email
        type: string
      - in: formData
        name: id
        type: integer
      - in: formData
        name: password
        type: string
      - in: formData
        name: userName
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResultResponse'
      summary: 注册用户
      tags:
      - 登录
  /v1/comments:
    get:
      description: 获取文章评论列表
      parameters:
      - description: 文章id
        in: query
        name: postId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResultResponse'
      security:
      - ApiKeyAuth: []
      summary: 获取文章评论列表
      tags:
      - 评论
    post:
      description: 创建评论
      parameters:
      - description: 新增的评论数据
        in: body
        name: CommentReq
        required: true
        schema:
          $ref: '#/definitions/request.CommentReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResultResponse'
      security:
      - ApiKeyAuth: []
      summary: 创建评论
      tags:
      - 评论
  /v1/posts:
    post:
      description: 用户创建文章
      parameters:
      - description: 用户创建文章的参数,只传content和title字段
        in: body
        name: PostDTO
        required: true
        schema:
          $ref: '#/definitions/dto.PostDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResultResponse'
      security:
      - ApiKeyAuth: []
      summary: 创建文章接口
      tags:
      - 文章
  /v1/posts/:id:
    delete:
      description: 删除文章
      parameters:
      - description: 文章id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResultResponse'
      security:
      - ApiKeyAuth: []
      summary: 删除文章
      tags:
      - 文章
  /v1/posts/detail:
    get:
      description: 获取文章详情
      parameters:
      - description: 文章id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResultResponse'
      security:
      - ApiKeyAuth: []
      summary: 获取文章详情
      tags:
      - 文章
  /v1/posts/list:
    get:
      description: 获取文章列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResultResponse'
      security:
      - ApiKeyAuth: []
      summary: 获取文章列表
      tags:
      - 文章
  /v1/posts/modify:
    put:
      description: 修改文章
      parameters:
      - description: 修改的文章数据
        in: body
        name: PostDTO
        required: true
        schema:
          $ref: '#/definitions/dto.PostDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResultResponse'
      security:
      - ApiKeyAuth: []
      summary: 修改文章
      tags:
      - 文章
  /v1/users/auth:
    get:
      description: 用户通过token和邮箱认证
      parameters:
      - description: 用户认证参数
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResultResponse'
      security:
      - ApiKeyAuth: []
      summary: 认证用户
      tags:
      - 用户
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"